* changes:
  Reduce a log message, stifle a gcc warning.

+++ vm/oo/Class.c
@@ -2418,7 +2418,7 @@ static void computeRefOffsets(ClassObject* clazz)
            * f->byteOffset is the offset from the beginning of
            * obj, not the offset into obj->instanceData.
            */
-          assert(f->byteOffset >= CLASS_SMALLEST_OFFSET);
+          assert(f->byteOffset >= (int) CLASS_SMALLEST_OFFSET);
           assert((f->byteOffset & (CLASS_OFFSET_ALIGNMENT - 1)) == 0);
           if (CLASS_CAN_ENCODE_OFFSET(f->byteOffset)) {
               u4 newBit = CLASS_BIT_FROM_OFFSET(f->byteOffset);
